MovieMatchDataAgg/requirements.txt
# This is an implicit value, here for clarity
--index-url https://pypi.python.org/simple/

mysql-connector-python==8.0.11
requests==2.21.0
bs4==0.0.1